+++ lib/grape-swagger/entity/attribute_parser.rb
@@ -13,7 +13,13 @@ module GrapeSwagger
 
         if entity_model
           name = endpoint.nil? ? entity_model.to_s.demodulize : endpoint.send(:expose_params_from_model, entity_model)
-          return entity_model_type(name, entity_options)
+
+          entity_model_type = entity_model_type(name, entity_options)
+          return entity_model_type unless documentation
+
+          add_array_documentation(entity_model_type, documentation) if documentation[:is_array]
+
+          entity_model_type
         else
           param = data_type_from(entity_options)
           return param unless documentation
@@ -25,7 +31,11 @@ module GrapeSwagger
             param[:enum] = values if values.is_a?(Array)
           end
 
-          param = { type: :array, items: param } if documentation[:is_array]
+          if documentation[:is_array]
+            param = { type: :array, items: param }
+            add_array_documentation(param, documentation)
+          end
+
           param
         end
       end
@@ -97,6 +107,12 @@ module GrapeSwagger
         return unless example
 
         attribute[:example] = example.is_a?(Proc) ? example.call : example
+      end
+
+      def add_array_documentation(param, documentation)
+        param[:minItems] = documentation[:min_items] if documentation.key?(:min_items)
+        param[:maxItems] = documentation[:max_items] if documentation.key?(:max_items)
+        param[:uniqueItems] = documentation[:unique_items] if documentation.key?(:unique_items)
       end
     end
   end